refactor(position-keeping): split position_repository.go by domain
Extract position CRUD operations to position_repository_position.go
and query/aggregation operations to position_repository_query.go.

position_repository.go retains the struct, constructor, shared helpers
(setSearchPath, scanPositions), errors, and interface assertion.

Remove from knownOversizedFiles as all files are now under 800 lines.
